Add HPCJobScheduler base and refactor
